Moved around some items, Sorted out better way to do devices, the bus, CPU and instructions.
Each instruction is now it's own sealed class inheriting either from the base instruction class or from a base class of the Instruction that fills out everything but addressing mode, from there each class then just defines which addressing mode it covers for the CPU.

The classes only have a single function, Execute which takes the current registers and bus, then uses the, to preform the instruction given.

This allows for individual testing of the instructions as well as allowing easy addition of more instructions in the future(The "Illegal" Opcodes) just by adding a class implementing the required information and function for the emulator to execute.

Also added a unit tests project for tests that will be created once there is enough of the code written out so that I can start adding instruction implementations and there proper tests along side them. allowing full and deep testing of the emulators every function and behavior to make sure it stays accurate to whatever machine is being emulated.

Currently working on getting 6502 working with RAM and then adding some NES related devices, such as cartridge, PPU, etc.

future plans are to add more devices, CPUs, etc.