Renamed ProjectFinished adding classes for all Legal Opcode on the 6502, Renamed Solution to EmulatorPrioject Renamed All Folders Relating to Emulator to match project names properly
Added some DebugDisplay Attributes to help with visualizing some of the data, so now instructions, the bus and registers are easier to read.

Split Console Application into separate project and made emulator a library so that it can be use in different projects later on.

next steps are Adding tests for each instruction that make sure they do what they are meant to and to implement each instruction as i complete the test for the instruction.