made some changes to use refs where i can to make it more readable that the register's instance is being modified when used.
added a couple of testing roms.

finished writing CPU class for now, reset,decompilation and a few new features have been added.

first new feature is for debugging, it now displays the registers, memory map, and other information in vs using the GetDebuggerDisplay Method. now when pausing or at a break point to can hover over the RAM,CPU and Registers to get a read out of current State. as a proof of concept is also prints out this information into the console app sample.

Started work on abstracting some of the CPU/device functions into interfaces for future works such as additional CPUs and devices, to make it easier to implement new ones in the future.